requirements.txt
# ── Core RAG ──────────────────────────────────────────────────────────────────
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# ── LLM Providers ─────────────────────────────────────────────────────────────
langchain-google-genai>=2.0.0      # Gemini 1.5 Flash  → export GOOGLE_API_KEY=...
langchain-openai>=0.2.0            # GPT-4o-mini       → export OPENAI_API_KEY=...

# ── Vector Store + Embeddings ─────────────────────────────────────────────────
chromadb>=0.5.0
sentence-transformers>=3.0.0       # Local embeddings (all-MiniLM-L6-v2)

# ── File Ingestion ─────────────────────────────────────────────────────────────
pandas>=2.0.0
openpyxl>=3.1.0                    # Excel (.xlsx / .xls)
pdfplumber>=0.11.0                 # PDF text extraction

# ── FastAPI Service ────────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
python-multipart>=0.0.9            # Required for file upload (FastAPI)


# ── Streamlit UI ───────────────────────────────────────────────────────────────
streamlit>=1.38.0

# ── Google SDK (Gemini) ────────────────────────────────────────────────────────
google-generativeai>=0.8.0
python-dotenv
langchain-classic
rank_bm25
langchain-chroma